web_backend/requirements.txt
flask
flask_cors
kafka-python
requests
browser_cookie3
psycopg2